requirements.txt
langchain"""Requirements for Local RAG application."""

langchain-ollama# Core dependencies

langchain-communitypython-dotenv>=1.0.0

langchain-text-splitterslangchain>=0.1.0

langchain-azure-ailangchain-community>=0.1.0

azure-cosmoslangchain-ollama>=0.1.0

azure-identitylangchain-text-splitters>=0.0.1

beautifulsoup4
# Azure Cosmos DB integration
azure-cosmos>=4.4.0

# LLM and embeddings
ollama>=0.1.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
requests>=2.31.0
pydantic>=2.0.0
